add server error

+++ site/views/layout.php
@@ -29,23 +29,19 @@
         include_once __DIR__.'/../'.'Helpers/SiteHelper.php';
 
         $domainId = SiteHelper::GetDomainId("cm93089_group5");
-
-        //$domains = RequestManager::GetAllDomains();
-        //print_r($domains);
-
-        //RequestManager::DeletePost($domainId,27);
-        //$posts = RequestManager::GetAllPost($domainId);
-        //$posts = json_decode($posts)->{'result'};
-        //print_r($posts->{'data'});
-        ?>
-
-        <div class="wrapper">
-        <?php
-        $posts = RequestManager::GetAllPost($domainId);
-        $posts = json_decode($posts)->{'result'};
-        foreach($posts->{'data'} as $value) {
-            $post = (array)$value;
-            echo '<div class="post">
+        if($domainId == null) {
+            echo '<div class="server-error">
+                    <div style="margin-bottom: 20px">К сожалению, сервер недоступен. Наши сотрудники уже решают эту проблему.</div>
+                    <img src="styles/img/fixies.png" height="400">
+                  </div>';
+        }
+        else {
+            echo '<div class="wrapper">';
+            $posts = RequestManager::GetAllPost($domainId);
+            $posts = json_decode($posts)->{'result'};
+            foreach($posts->{'data'} as $value) {
+                $post = (array)$value;
+                echo '<div class="post">
                 <a class="delete" title="Delete" href="views/deletePost.php?id='.$post["id"].'"></a>
                 <a class="link" href="'.$post["url"].'">
                 <div class="post-title">'.$post["title"].'</div>
@@ -76,11 +72,10 @@
 					
 					<button class="btn btn-primary btn-custom" href="views/updatePost.php" name="submit" type="submit">Update</button>
 					</div></form>';
+            }
+            echo '</div>';
         }
-        //$post = RequestManager::CreatePost($domainId);
-        //print_r($post);
         ?>
-        </div>
 
         <?php
         echo '<form name="insert" action="views/addPost.php" method="post">
